requirements.txt
maim_message
websockets
asyncio
python-dotenv
wxauto